+++ lib/app.js
@@ -1,11 +1,11 @@
 const express = require('express');
 const app = express();
 const morgan = require('morgan');
-const auth = require('./routes/auth/auth');
-const twitch = require('./routes/twitch/twitch'); 
-const favorites = require('./routes/favorites/favorites'); 
-const ratings = require('./routes/ratings/ratings');
-const community = require('./routes/community/community'); 
+const auth = require('./routes/auth');
+const twitch = require('./routes/twitch'); 
+const favorites = require('./routes/favorites'); 
+const ratings = require('./routes/ratings');
+const community = require('./routes/community'); 
 const jwt = require('./jwt');
 
 app.use(morgan('dev'));
@@ -36,7 +36,8 @@ function checkAuth(req, res, next) {
 app.use('/api/auth', auth);
 app.use('/api/twitch', checkAuth, twitch);
 app.use('/api/favorites', checkAuth, favorites);
-app.use('/api/community', checkAuth, community);
+// Just use favorites
+// app.use('/api/community', checkAuth, community);
 app.use('/api/ratings', checkAuth, ratings);
 
 module.exports = app;
