+++ lib/routes/community.js
@@ -1,6 +1,7 @@
 const router = require('express').Router(); 
-const client = require('../../db-client');
+const client = require('../db-client');
 
+// This is unnecessary, same as /api/favorites!
 router.get('/', (req, res) => {
   client.query(`
     SELECT user_name
