+++ lib/routes/favorites.js
@@ -1,6 +1,6 @@
 /*eslint-disable-next-line*/
 const router = require('express').Router(); 
-const client = require('../../db-client');
+const client = require('../db-client');
 
 router.get('/', (req, res) => {
   
@@ -21,14 +21,21 @@ router.get('/', (req, res) => {
     const body = req.body;
 
     client.query(`
-      SELECT profile_id
+      SELECT 
+        id, 
+        profile_id as "profileId",
+        user_name as "streamer" -- assuming you made this changegetRandomScore()
       FROM favorite
       WHERE user_name = $1;
   `,
     [body.user_name])  
       .then(result => {
         if(result.rows.length > 0) {
-          res.status(400).json({ error: 'streamer already favorited' });
+          // Doesn't need to be error, objective is already satisfied:
+          // res.status(400).json({ error: 'streamer already favorited' });
+
+          // so just send back the data:
+          res.json(result.rows[0]);
           return;
         }
 
@@ -40,8 +47,8 @@ router.get('/', (req, res) => {
           VALUES ($1, $2)
           RETURNING
             id,
-            user_name,
-            profile_id;
+            user_name as "userName",
+            profile_id as "profileId";
 `,
         [body.user_name, req.userId])
           .then(result => {
@@ -50,12 +57,14 @@ router.get('/', (req, res) => {
       });
   })
 
+  // oh no! this would delete all rows with that stream name!
   .delete('/:name', (req, res) => {
     client.query(`
       DELETE FROM favorite 
-      WHERE user_name = $1;
+      WHERE user_name = $1
+      AND profile_id = $2; -- needs to be this user!
   `,
-    [req.params.name])
+    [req.params.name, req.userId])
       .then(result => {
         console.log('DELETE FINISH');
         res.json({ removed: result.rowCount === 1 });
