+++ lib/routes/twitch.js
@@ -6,7 +6,6 @@ router.get('/', (req, res) => {
     .set('Content-Type', 'application/json')
     .set('Client-ID', '8sb2kt99biht5q3k79k7hsejyj0q2y')
     .then(response => {
-      console.log('response', response.body);
       const twitchResults = response.body.data.map(stream => {
         return {
           user_name: stream.user_name,
