+++ package.json
@@ -24,7 +24,6 @@
     "morgan": "^1.9.1",
     "nodemon": "^1.18.9",
     "pg": "^7.7.1",
-    "router": "^1.3.3",
     "superagent": "^4.0.0"
   },
   "devDependencies": {
