+++ scripts/create-tables.js
@@ -5,7 +5,7 @@ client.query(`
 
   CREATE TABLE IF NOT EXISTS profile (
     id SERIAL PRIMARY KEY,
-    username VARCHAR(256) NOT NULL,
+    username VARCHAR(256) NOT NULL UNIQUE, -- enforce unique at a db level
     first_name VARCHAR(256),
     email VARCHAR(256),
     hash VARCHAR(256) NOT NULL
@@ -13,13 +13,15 @@ client.query(`
   
   CREATE TABLE IF NOT EXISTS favorite (
     id SERIAL PRIMARY KEY,
-    user_name VARCHAR(256) NOT NULL,
+    streamer_name VARCHAR(256) NOT NULL,
     profile_id INTEGER NOT NULL REFERENCES profile(id)
   );
-  
+    
+  -- why is profile_id not included?
   CREATE TABLE IF NOT EXISTS rating (
     id SERIAL PRIMARY KEY,
-    user_name VARCHAR(256) NOT NULL,
+    profile_id INTEGER NOT NULL REFERENCES profile(id),
+    streamer_name VARCHAR(256) NOT NULL,
     score INTEGER NOT NULL
   );
 `)
